+++ src/components/features/protocol/data-models/view/layout.tsx
@@ -5,7 +5,7 @@ import { useMemo, useState } from 'react';
 import { TabPanel } from '@/components/atoms/tabs';
 import { HeadContainer } from '@/components/molecules/head-container';
 import ModalRight from '@/components/molecules/modal/modal-right';
-import ConfirmDialog from '@/components/organisms/confirm-dialog/confirm-dialog';
+import ConfirmDialog from '@/components/molecules/modal/confirm-dialog';
 import { ClientNav } from '@/components/organisms/navbar/client-nav';
 import { DashboardTemplate } from '@/components/templates/dashboard';
 import { query } from '@/constants/queries';
